Install Gos Python package

This function wraps installation functions from reticulate to install the Python package gos.

install_gosling(
  method = c("conda", "virtualenv"),
  envname = "r-reticulate",
  version = NULL,
  ...
)

Arguments

method

character, indicates to use "conda" or "virtualenv"

envname

character, name of environment into which to install

version

character, version of Gos to install. For general use of this package, this is set automatically, so you should not need to specify this.

...

other arguments sent to reticulate::py_install()

Value

invisible NULL, called for side-effects

Details

This package uses the reticulate package to make an interface with the Gos Python package. To promote consistency in usage of reticulate among different R packages, it is recommended to use a common Python environment, called "r-reticulate".

Depending on your setup, you can create this environment using reticulate::conda_create() or reticulate::virtualenv_create(), as described in this reticulate article.